Rename mypyc.ops_primitive to mypyc.primitives.registry
