+++ mypyc/irbuild/builder.py
@@ -40,7 +40,7 @@ from mypyc.ops import (
     is_none_rprimitive, object_rprimitive, PrimitiveOp, OpDescription,
     is_object_rprimitive, FuncSignature, NAMESPACE_MODULE, RaiseStandardError, FuncDecl
 )
-from mypyc.ops_primitive import func_ops
+from mypyc.primitives.registry import func_ops
 from mypyc.primitives.list_ops import list_append_op, list_len_op, new_list_op, to_list, list_pop_last
 from mypyc.primitives.dict_ops import dict_get_item_op, dict_set_item_op
 from mypyc.primitives.misc_ops import (
